autopkgtest: Save pending.json after every test request
We had a case where the autopkgtest queues DoSed because britney was
crashing after requesting each reverse dependency for a perl upload, but
before it had written pending.json out so it knew what not to request
again.  This was 25,000 requests per arch...

Let's write pending.json straight after sending each request, so that
the next run - even after a crash - won't re-request the same things
again.

Ideally of course we would remain crash free, but this is a mitigation
against bugs.