autopkgtest: Re-use amqp file handle between excuses and open 'w'
Currently if britney crashes, we might be left with a half-written amqp
submission file. We may then go on to write the same requests into it on
the next run, when we process the excuses for a second time.

This can happen since we open the file for each excuse in 'a'ppend mode.
If we instead open in the initialiser in 'w' mode, not only will we
always start with an empty file on each run but it'll also be a bit more
efficient.