suiteloader: non-free-firmware was added to the archive
Today, unstable and testing have non-free-firmware added to their components, but (oldstable-)proposed-updates
and (old)stable not. britney currently assumes the components are the same across the board and takes the
components from the target suite. This commit is not a proper fix, but works around the issue by not failing
on not-found files for the non-free-firmware component.