Optimize out deadcode that triggers maximum recursion error
The `get_auto_hinter_hints` method uses `find_related` to find
candidate hints. The code looked like this:

```python
    hint = find_related(e, set(), True)
    if isinstance(hint, dict) and e in hint:
       ...  # Apply hint
```

This implies that we only use `hint` (i.e., the return value of
`find_related`) when `find_related` returns a dict.  Looking at
`find_related`, we see that it always returns *either* its second
argument or `False`.  Given the second argument is always a `set`, we
can conclude that the condition for above `if`-statement is always
False and we can optimize the code by removing the if + the call to
`find_related`.

As a side-effect, we avoid the stack overflow / maximum recursion
error triggered by find_related!

Closes: #1010660
Signed-off-by: Niels Thykier <niels@thykier.net>