Make policy loading more declarative
Turn a series of "if (optionally) + load policy" statements into a
list of `PolicyLoadRequest.always_load` vs.
`PolicyLoadRequest.conditionally_load` entries.

I feel this puts more emphasis on the important aspects - such as the
order of the policies and when they are conditionally loaded, what is
the configuration option + default for that policy.

Additionally, it is now a one-line change to turn an unconditional
policy into a conditional one (just define the config name and the
default).

Signed-off-by: Niels Thykier <niels@thykier.net>