Set mindays to 2 days (was zero)
Downside: it will slow down a bit the migration of packages from
kali-dev to kali-rolling. So far, the only slow down that I know of was
due to tests (britney would wait for all test to complete, which might
or might not be quick). When there was no test, package would go
straight from kali-dev to kali-rolling, without any waiting time.

So with this commit, things will be a bit less "instant", hopefully that
will be fine and we won't even notice.

Upside now: with this delay, we will have time to notice breakage that
is caught by the daily QA on kali-dev, so we can block bogus packages
before they reach kali-rolling.

Context: lately we had a breakage due to a conflict between desktop-base
and kali-themes, it was quite bad, it caused apt to fail. The two
packages tried to install the same file, and this is not something that
is caught by britney. However our daily QA on kali-dev caught it.

2 days is quite short to react on kali-dev's daily QA, we might want to
increase this delay in the future, depends on how well it works in
practice.