autopkgtest: skip tests on archs where source is not build
Until this commit, the autopkgtest policy would trigger a test on
every architecture for mixed arch:all + arch:<something> sources. Now,
an arch will be skipped if no binaries are built on this arch (but
there are binaries built on other architectures).

A nice catch is that also arch:<something> with arch:all reverse test
dependencies will not be scheduled in vain on archs where they are not
build.