Make kali-run-britney.sh simply call run-standalone.sh
With this commit, we basically drop our scripts kali-run-britney.sh and
kali-prepare-data.sh, in favor of using run-standalone.sh.

To clarify, the script kali-run-britney.sh still exists, but all it does
is call run-standalone.sh and store/rotate the logs.

Refer to upstream commit 0f75722311972420bbd9fa95c7cb077acb93ccbd from
sdelafond: this is when run-standalone.sh was introduced, along with
other scripts. In particular, scripts/debci-put.py used to be
Kali-specific, it was reworked to be more generic and is now provided by
upstream. Similarly, run-standalone.sh is a more generic approach that
does what kali-run-britney.sh and kali-prepare-data.sh used to do for
Kali.

With all that being said, run-standalone.sh is not exactly on par with
what we had in the Kali scripts, and not completely flexible enough, so
we still need to patch it a bit for our use-case.

Finally: note that, with this change, we lose support for the
--no-prepare argument in the script kali-run-britney.sh, however:
- I'm don't know what was the use-case for that
- I don't know if anyone used it
- All it did was invoke britney only (a one-liner), so anyone can still
  do that if they wish.

So if there's a use-case X, we can document "in order to do X, you can
just run britney as such: xxxxx"